requirements.txt
pandas>=1.3.0
pymysql>=1.0.2
psycopg2-binary>=2.9.0
pyodbc>=4.0.32
sqlalchemy>=1.4.0
geopandas>=0.12.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.4.0
geoalchemy2>=0.12.0
matplotlib>=3.4.0
pillow>=8.0.0